[FIX] hr_appraisal: fix test_appraisal_next_appraisal_date_uppcoming_appraisal
In test_appraisal_next_appraisal_date_uppcoming_appraisal, a search is done the employees to
check the appraisal date based on the number of appraisal the employee had.

If the appraisal count is 0, the domain checks if the create_date of the employee is within the month
of the duration_after_recruitment.
Since a relativedelta is applied to the current date, if the current date has 30 days and the create_date is
on the 31st of a month, it won't find the employee.

Instead, the create_date is checked against a date range starting from the first of the month to the last day

task-2726422

closes odoo/enterprise#23225

X-original-commit: 10cb6cc9b74bcd1f5659ac92910346fcf1122720
Signed-off-by: Kevin Baptiste <kba@odoo.com>