[FIX] account_reports: fix unfold bug in financial reports
Currently, report lines in the BS, P&L and General Ledger that are unfoldable,
need to be clicked twice in order to unfold (after scrolling).
This is not supposed to be the case.

This happens because a clone of the report table is made in order
to improve the management of the width of columns with ellispis in the reports.

task-2711352

closes odoo/enterprise#23248

X-original-commit: f6cd59e23a27bc6c1c58be33b17bba34626293e4
Signed-off-by: William André (wan) <wan@odoo.com>