[FW][IMP] appointment: add performance tests
Purpose is to add performance tests for appointment module in order to be able
to quantify improvements of incoming code cleaning. Tests are added currently
with working hours check, on both type: either with a list of employees (type
custom) or with a fixed employee (type work_hours).

When forward porting to master it will be interesting to check effect of work
hours on overall computation and performances as it has been split from base
appointment in 15.2+ .

Counters show we have some room for improvements to say the least.

Task-2766574 (Appointment: improve performance)
Task-2631088 (Appointment: profile and add performance tests)

X-original-commit: odoo/odoo@3fa07d5de523fe54d34b7e3f0b7848032fe99bff
Part-of: odoo/enterprise#24559