[IMP] helpdesk: use the average of rating value for satisfaction
Before this commit, we display an emoji to show if the customer is
satisfied, okay or not for the ticket via the last rating value. The
problem is we use only the value for the last rating for each ticket.

This commit uses all ratings for each by using the average of rating
value for each ticket to know if all the customers that done a rating are
satisfied or not.
Moreover, we add 3 filters in the helpdesk.ticket model using the average
of rating values:

1. Satisfied => rating_avg >= 3.66
2. Okay => 2.33 >= rating_avg < 3.66
3. Dissatisfied => rating_avg < 2.33

task-2671848

Community PR: odoo/odoo#81027

Closes #22773