[FIX] industry_fsm_sale: product quantity validation error in Android
Steps to reproduce:
- Go to Field Service App from an Android device
- Create a new task
- Select a title and a customer
- Click on the smart button "Products"
- Add more than 9 characters or some alphabetical characters

Result:
The alphabetical characters are visible and is possible to add more than
9 characters. The validation is not performed.

Expected result:
As on a desktop browser it shouldn't be possible to add more than 9
digits and the shown number should be a natural (no alphabetical
characters or special characters).

Keypress events are not working on Android and also the
KeyboardEvent from an Android Virtual Keyboard doesn't return the
pressed key identifier (keyCode).
I switched the span contenteditable with a simpler input element. Now
the UI doesn't prevent the insertion of an invalid number, but in this
case the value is highlighted by a red color. The server will check and
save the new value only if valid.

opw-2710590

closes odoo/enterprise#24189

X-original-commit: e7fdfb102dad3fb15c3c5cc2ff51eec9540f290d
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>