[FIX] account_intrastat: Fix report country code function
The function _get_report_country_code is inherited from account.report
on the account.intrastat.report model, but returns None in this context.
This creates problems in the child modules (such as in l10n_nl_intrastat),
where the function is used to determine whether a particular export option
for the report is provided to the user). This fix defines the function
_get_report_country_code on the account.intrastat.report model, so that a
usable country code is returned.

closes odoo/enterprise#23356

Task-id: 2728182
X-original-commit: 566f35bb83815cb1edd2b0b77da1d4166c559373
Signed-off-by: Olivier Colson <oco@odoo.com>