[IMP] marketing_automation: remove the elements related to "contact_list_ids" in the mailing form view
Purpose
=======
A button has been added next to `contact_list_ids` in the mailing form
view. So now, we remove the parent element (so the label, the field and
the button) in the view defined in marketing automation.

Those elements are not useful for the "mailings used in marketing
automation".

Task-2703521

closes odoo/enterprise#23263

Related: odoo/odoo#82333
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>