[IMP] web_enterprise: HomeMenu use CommandPalette
Currently, there are two implementations that allow you to search and
navigate through apps/menus. The searchBar of the homeMenu and the "/"
namespace of the command palette.

This commit aims to unify the two implementations. When a user performs
a search in the homeMenu, the command palette will open and the search
will be performed in it.

We are taking the opportunity of this commit to use the new hotkey
system in the homeMenu.

closes odoo/enterprise#23285

Task-id: 2730493
Related: odoo/odoo#82629
Signed-off-by: Géry Debongnie <ged@odoo.com>