[MOV] {test_}web_gantt,project_enterprise: bring dependency in web_gantt
This commit moves the dependency mecanism that had been developed in
project_enterprise into web_gantt.

This commit creates a basic rescheduling behavior in web_gantt that can
be extended in other modules (as it is done for project_enterprise in
this commit).

An additional module `test_web_gantt` has been created in order to test
the the rescheduling feature on some fake models.

Functionaly wise, we reduced the scope of the rescheduling algorithm, in
order to prevent undesired results. This commit prevents rescheduling
so called `loops` when it detects that it can not solve them. `Loops`
are not cyclic dependencies but records that are linked together, forming
a loop, as illustrated here below:

      ------> 4
     |        |
     |        v
     1 -> 2 ->3

task-2691454
Closes #23370

Related PR: odoo/odoo#85395