[FIX] l10n_be_hr_payroll: Fix field called from other module
Initial behavior:
When trying to generate a payslip with the structure cp200 - Termination fees,
an error occurs if the module l10n_be_hr_contract_salary is not installed.

After this commit:
It is possible to create the payslip with that structure without the module previously mentionned.
If the module is installed, the original computation will be applied.

closes odoo/enterprise#23668

Taskid: 2639082
X-original-commit: 99ef16bf0492cc960718d6be1283079aa23c598f
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>