[FIX] web_studio: Access to reports when no data available
Before this commit, when you try to open the report "Quotation/Order"
in studio when there is no record, there is a server error.

This commit fix it by raising a user error when he tried to open an
empty report in studio.

Steps to reproduce:
==================
1) Use a DB without demo data
2) install sales ad studio
3) go on sales -> studio
4) click on Reports
5) try to open any reports

Expected Behaviour:
==================
We should raise a user error to explain what append to the user

Actual Behaviour:
================
There is a server error and a traceback

closes odoo/enterprise#23334

X-original-commit: 3539976a060b855148c28aeff7e3c8abb3285983
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Bruno-brsy <brsy@odoo.com>