[FIX] web_studio: fix indeterministic crash in _updateSidebar
Previously, calling _updateSidebar twice in quick succession could
result in a crash. This was caused by the fact that the sidebar is
instantiated and stored on "this", then it is appended to a fragment,
which is asynchronous. After the sidebar has been appended to the
fragment, the code attempts to append it to the document, but it does so
by reading "this.sidebar" which may not be the same sidebar that was
appended if another call to _updateSidebar was madei, causing the
assumption that the sidebar has an $el to be wrong.

This commit fixes that by capturing the sidebar that was actually
appended in the closure of the continuation, and uses it instead of
going through self.

closes odoo/enterprise#23371

X-original-commit: 425bab5ef662b0e07652c48da83d29a2880392ef
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Samuel Degueldre <sad@odoo.com>