[FIX] appointment: add a sequence to question in demo data
As calendar.appointment.question has no default sequence value,
the database creates the demo data question "Symptoms" with NULL
sequence number.

This causes the ordering in front end issues as this question will
always appear at the end of the list, as long as handle widget has
not been used back end to give the sequence number a value.

Task-2621327

Part-of: odoo/enterprise#20433