[FIX] planning: fix recursive create resource and employee
Before this commit, we could have a recursion when creating a resource and
an employee because of the `from_planning` key in the context. When
the user creates a resource via the form view of a `planning.slot`, by
default, we want to create an employee resource. So, in the create
method, we check if the `from_planning` key is in the context. If yes then
we create an employee. The problem arise when the employee is created as the
employee creates a `mail.thread` and then creates a resource. In that case,
if the flow come from planning, the context key remains and so the resource
will create again the employee, causing a recursive loop.

This commit removes the `from_planning` context key when the employee is
created in the create method in the `resource.resource` model.

closes odoo/enterprise#23336

X-original-commit: a8e79c2fbeb37cb34ab504cb0ed3ee1f2fca0652
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>