[IMP] delivery_dhl: demo data - credentials for US
Since the last update, there were no more credentials for US shipping in
the demo data. Since this is not very practical, here are the
credentials for US shipping.

Since the US account number was requested from the same account as the
EU account number, the SiteID as well as the password are the same for
the 2 account numbers.

These credentials are demo data and need to be considered as such. this
is why they are no longer in the data directory, but in the dedicated
demo directory. The packages data on the other hand are really data.

closes odoo/enterprise#23594

X-original-commit: 64bc01ebd1cd4a87590e0e6f3fde0a4b1b23d27a
Signed-off-by: Arnold Moyaux <arm@odoo.com>