[REF] web_enterprise: stop using component ref in navbar
Previously, the ref to the menu brand element was in the community
template but unused in community. Since it is a component ref it will no
longer be supported in owl2 so it has been removed. Ideally, we would
like to simply wrap the corresponding component in a t-if and not render
it when it is unnecessary, unfortunately, since we want to remove or
display this element in a synchronized manner with the mounting of the
home-menu action, we cannot afford to rerender the navbar as this will
cause flickering and a possibly inconsistent state between the navbar
and the action container. Because of that, we have to toggle a class
that will hide the menu brand imperatively so that it is synchronous.
Because refs on components will no longer be valid, this is now done
using a querySelector on the element's content.

This commit also refactors the onMounted and onPatched calls to use
useBus and useEffect instead.

closes odoo/enterprise#23289

Related: odoo/odoo#82326
Signed-off-by: Géry Debongnie <ged@odoo.com>