[IMP] hr_expense_extract: improve UX (back2basics)
Community PR - odoo/odoo#81904
Upgrade PR - odoo/upgrade#3206

This commit is addressing the following:

we used to log a message when we attached the receipt to the expense form.
The message is redundant with the overview available in the chatter and
the files attached to the record. Thus we removed it.

We show the expenses that are scanned from the list view (before they
are processed by OCR and fields updated) at the top of the list - muted.

Now we trigger cron (that updates expense fields that are processed by OCR)
'immediately' after uploading attachments.

Before, in case we had multiple active currencies and OCR didn't manage
to recognize the currency, we got an error. Now it is fixed.

task - 2687999

closes odoo/enterprise#23980

Signed-off-by: Kevin Baptiste <kba@odoo.com>