[FIX] mail_enterprise: fix code of message swiper
Avoid hack because of slot (as strongly suggested on the original PR it should
not have used slot in the first place, but out of scope to remove it).

Indeed lifecycle hooks and render overrides are both broken when using slots,
and shouldUpdate always true on message is not acceptable until Owl 2.

task-2728107

closes odoo/enterprise#23331

X-original-commit: e22b741a8fed22182c8e2a69d933142a1bec56e5
Related: odoo/odoo#82415
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>