[FIX] web_dashboard: traceback on date aggregate with a falsy value
Having a date/datetime field aggregate with a falsy value in the read_group led to a traceback because it had a fallback value of `0` and this is not supported by the date/datetime formatter.
After this commit, it falls back to `NaN` and the date/datetime formatters are therefore not called at all in this case.

opw-2698505

X-original-commit: 50f0286c9b317e30d5f76df4c097c9cfa17ba7d8
Part-of: odoo/enterprise#23421