[REF] delivery_easypost: Packages/Commodities in base
Each delivery provider, requires some information about the packages
that need to be sent as well as about the commodities (for commecial
invoices). Since this is needed by each provider, and the required
values are very similar, these packages and commodities can be done one
step ahead: in the base delivery module.

reference-branch=
https://github.com/odoo-dev/odoo/tree/master-ref_ups_add_insurance_percentage-goa

Part-of: odoo/enterprise#22907