[FIX] l10n_be_hr_payroll: Break down representation fees
Purpose
=======

From the 1rst of January 2022, a distinction should be made between the
employer expenses based with or without serious standards.

The representation fees are based on serious standard until 279.31 €

Only part of the representation costs are pro-rated because certain costs are fully
covered for the company (teleworking costs, mobile phone, internet, etc., namely:
 - 144.31 € (Tax, since 2021 - coronavirus)
 - 30 € (internet)
 - 25 € (phone)
 - 80 € (car management fees)
 = Total € 279.31

The remaining part of the representation fees should be declared separately
in the payslips and the 281.10 sheet.

Taskid: 2732895
X-original-commit: 431bf4c4207538ab643b4e017739bf92a0f0a277
Part-of: odoo/enterprise#23457