[IMP] sale_renting*: adapt the renting modules to the sale_temporal logic
Before this commit, sale_renting had his own logic to handle the sale of "periods" based on product pricings.
This commit aims to rely on the sale_temporal module that handle the same logic but make it overridable by other modules.
The functionnalities of sale_renting do not change here but the sale_temporal module makes them more generic.

taskid: 2568007