[IMP] l10n_it_reports : BS and P&L for Italy
This change will modifiy the l10n_it_reports module by adding the data folder and the
files for the balance sheet and the profit and loss report. Those reports follows the
model presented by the Italian Civil Code as of December 2020. It also
add a reduce version of the balance sheet that can be use by smaller
companies.

These changes have been made in order to further complete the Italian localization.

task-2468834

closes odoo/enterprise#24498

X-original-commit: c6521280e5f2debb90eb4dcdb8872aa449be8107
Related: odoo/odoo#84857
Signed-off-by: Olivier Colson <oco@odoo.com>