[FIX] l10n_pe_edi: Fix unit tests.
At the moment, if you run the tests in test_edi_iap.py,
test_edi_digiflow.py and test_edi_sunat.py, they will fail because
the OSE sends an error response:

3014|El codigo de leyenda no debe repetirse en el comprobante.
- Detalle: xxx.xxx.xxx ticket : 20220000000000101174262 error:
Leyenda : 1000: 3014 (nodo: &quot;cbc:Note/languageLocaleID&quot;
 valor: &quot;1000&quot;

This error is caused by the presence of two  `<cbc:Note>` nodes with
the same `languageLocaleID` attribute value in the XML that we send
to the OSE.

This is because we set the l10n_pe_edi_legend field to '1000' in the
test invoice. As a result, in the generated XML, there are two
`<cbc:Note>` nodes with `languageLocaleId="1000"`.

This is because of
https://github.com/odoo/enterprise/blob/567f476ae7d82bdda84c84f8b73cd17775e24df7/l10n_pe_edi/data/2.1/edi_invoice.xml#L27
introduced in commit
https://github.com/odoo/enterprise/commit/567f476ae7d82bdda84c84f8b73cd17775e24df7

If we fix this first error, we still get another error because the
invoice is too old (its date is 01 Jan 2017).

Changing the invoice date to the current day causes the invoice to be
accepted by the OSE.

Additional fix for v15
Because the EDI test mode was removed in v15 - see odoo/odoo#71650
the tests need to be edited to not call
AccountEdiTestCommon._process_documents_web_services, otherwise the
invoice will get sent twice.

2nd additional fix for v15
The AccountEdiFormat._l10n_pe_edi_sign_invoices_sunat_digiflow_common
method didn't return {'success': True} on success. That is now fixed.

closes odoo/enterprise#27563

X-original-commit: 1c8e5467cc43912373571118f0959c476ce9a1e7
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>