[REM]: stock_barcode_picking_batch: Remove auto-batch
As a new auto-batch feature is made in the core stock module, the
auto-batch feature exclusive to barcode is no longer necessary.
Since if auto-batches are enabled on the picking type level, pickings
will be already in batches by the time they are created, so no manual
intervention should be required to batch them anymore.

Task-2670580

closes odoo/enterprise#23484

Related: odoo/odoo#81533
Related: odoo/upgrade#3205
Signed-off-by: Arnold Moyaux <arm@odoo.com>