[FIX] l10n_mx_edi(_stock): cancellation flow update
We needed to update the cancellation API to the 3 PACs, because
depending on the cancellation reason 01 or 02 you indicate
if the invoice you are cancelling has been substituted by another or not.

https://solucionfactible.com/sfic/capitulos/timbrado/ws-timbrado-cancelar.jsp
https://developers.sw.com.mx/knowledge-base/cancelacion-cfdi/
https://wiki.finkok.com/doku.php?id=cancelcfdi

It means we need to add an optional param in the cancellation method as well
with the invoice we substitute (or none of we don't substitute).

The flow can also change a bit.  So you would send invoice1 and then invoice2 with
04|folioinvoice1 as cfdi origin and then only you would cancel invoice1.

We added a computed field that calculates the invoice that replaces the one you
are looking at.

This is done for both pickings and invoices.

We got plenty of opw for the cancellation flow failing because
we did not send the cancellation reason.

closes odoo/enterprise#23590

X-original-commit: ccdd74cff898cb6cf1e956c34fcebd10276e663d
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Josse Colpaert <jco@odoo.com>