[IMP] account_asset: make original_value editable
The goal is to allow the user to edit an asset's
original value, even when there are original_move_line_ids.

We don't want to restrict the user from doing what he/she wants,
but a non-blocking  warning will pop up if he/she sets a value
that differs from the original_move_line_ids.

Part-of: odoo/enterprise#20950