[IMP] account_asset: improve Fixed Asset Account management
The goal of this commit is two fold:
- Make sure that when you create an asset model from the CoA,
it uses the current account as the Fixed Asset Account (and make
the field invisible so it cannot be changed).

- Make it possible to create an asset model without a Fixed Asset Account,
so that it can be used with multiple accounts (i.e. create an asset model
for cars and be able to apply it on sub accounts, like "VW cars",
"Opel cars"...).

Part-of: odoo/enterprise#20950