[FIX] sale_amazon, easypost_delivery: rely on delivery type
A modification made by Amazon on October 1st restricts the allowed
values for the carrier name. Since it now expects it in a formatted
state, this commit will allow Odoo to send the delivery type of the
carrier rather than its name, which can be customized by the user.

opw-2665099

closes odoo/enterprise#24204

X-original-commit: 310cea01b7e7ac8bafdde5e3108e2fc494c4feb4
Related: odoo/odoo#84291
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>