[FIX] delivery_dhl: fix DHL shipment validation
Steps to reproduce the bug:
- Install delivery_dhl
- Go to shipping Methods > click on “DHL” > Enable “Dutiable Material”
- Create a SO:
    - Add any customer
    - Add any storable product
    - Click on “Add Shipping” and select “DHL”
- Confirm the SO
- Try to validate the delivery

Problem:
A traceback is triggered `”Missing element TermsOfTrade (ShipmentRequest.Dutiable.TermsOfTrade)”`

Since the migration of the DHL api from version 6.3 to 10. The `"dutiable.termsOfTrade"` field has become required.
The field was added to `dct_dutiablle` instead of `dutiable` in this commit: https://github.com/odoo/enterprise/commit/abf8955f24f252e88aa93793180f6593e27275df

We should use the incoterm defined in the SO linked to the delivery or the default incoterm of the current company.
For more information, please refer to: https://mydhl.express.dhl/ca/en/help-and-support/customs-clearance-advice/customs-terms-of-trade.html

opw-2725851
opw-2730743

closes odoo/enterprise#23532

X-original-commit: 13bb8d3470d664e991de3e1da406fe22bfbb0e0a
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>