[FIX] sign: fix private signers
before this commit, if one of the signer is a private partner.
A normal user will get many access errors

The following scenario is not supported
'mail.thread' will ignore private partners added to the follower list without raising error
But if a contact partner becomes a private partner after added to a follower list.
Normal users without permission to access private partners will get an access error popover.
But the main business in eSign will not be blocked

Taskid: 2659342
Part-of: odoo/enterprise#22813