[FIX] l10n_de_pos{_res}_cert: error handling when sending to Fiskaly
Before this commit, the user was blocked if there was a network issue and he couldn't send his requests to Fiskaly. He couldn't proceed and get a non-compliant receipt.

This commit adapts the code after the changes made by https://github.com/odoo/odoo/pull/73725 in order to properly handle the error when sending requests to Fiskaly. The user is no longer blocked if there's a connection issue.

closes odoo/enterprise#23697

X-original-commit: 7bc4b3de78f8510db15c1f59b18748a192aa5dd9
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>