[FIX] spreadsheet_document: escape filter name
before this commit,
if we use special characters* like (,[  in filter name it will give trace back
while renaming. this is happening because we use regular expression and it
consider special character like (,[ as part of regular expression in replace
method

after this commit,
f we use special characters* like (,[  in filter name it will not give trace back
while renaming. to solve this issue escape all special character in label.

*: any escape string - See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions

task-2691559

closes odoo/enterprise#24896

X-original-commit: ffc9567cc4d31e1707f6939b6f98edbe2204a5b7
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Co-authored-by: Rémi Rahir <rar@odoo.com>