[FIX] appointment: fix the share availabilities feature
When in year mode, it was not possible to get the link if there
was not a slot created before in another mode.
Fix also the behavior in the year mode when selecting a date.
Indeed, selecting a date in year mode create an "allDay" slot.
If the user select the same date or a date containing events, it
should remove only the "allDay" slot linked to the date.
We don't want to remove all events of the date to avoid error of
the user.

task-2738045

closes odoo/enterprise#24719

X-original-commit: 31da5ceab89f760cd3e6a6729b65577def13a69f
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>