[FIX] sale_subscription: avoid infinite loop when generating invoice
How to generate ?

- Install subscriptions plugin
- Create a new subscription with the following parameter:
   -> to_renew must be checked
   -> the template should "Send after successful payment"
   -> add a product
- Change the subscription status to "In Progress" and change
  the start date and the date of next invoice in the past.
- Duplicate 20 times the subscription (create an action if necessary)
- Start the cron tast named 'Sale Subscription: generate recurring
invoices and payments'

The bug :

If you have more than 20 subscriptions for which an invoice must be
generated while it's not possible, the cron task will loop infinitely.
This is due to the fact that when there are more than 20 subscriptions,
the cron task will be triggered again but the same 20 subscriptions will
be fetched, etc.

opw-2725806

closes odoo/enterprise#24102

X-original-commit: bb6176bbe5b768767cbdd654c39c348c6afdec6c
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>