[IMP] delivery_easypost: add shipping insurance
Shipping insurance is a service which may reimburse senders whose parcels are
lost, stolen, and/or damaged in transit.

With easypost, we need to validate the order first, then call the `insure`
endpoint on each shipment.

Insurance value is in USD, so it needs to be converted afterward. To
calculate the insurance cost, easypost uses the insurance_fee_rate and
the insurance_fee_minimum, based on the connected user. These values
need to be retrieved in order to get a rate that takes the insurance
into account.

task-2701459
PR in base: https://github.com/odoo/odoo/pull/83896

closes odoo/enterprise#22907

Signed-off-by: Arnold Moyaux <arm@odoo.com>