[IMP] delivery_dhl: add shipping insurance
Shipping insurance is a service which may reimburse senders whose parcels are
lost, stolen, and/or damaged in transit. With DHL, we declare an Insured value
for all the packages at once. This value corresponds to the cost of the products
in the packages multiplied by the percentage of the shipping insurance.

For convenience, and because it is an optional field according to the
documentation (and verified by tests), the BillingCode in
`ship-val-global-res-10.0.xsd` was set as optional:
```diff
- <xsd:element name="BillingCode" type="dhl:BillCode">
+ <xsd:element name="BillingCode" type="dhl:BillCode" minOccurs="0">
```

task-2701459
PR in base: https://github.com/odoo/odoo/pull/83896

closes odoo/enterprise#23124

Signed-off-by: Arnold Moyaux <arm@odoo.com>