[FIX] account_reports: Fix tax report base amounts with refund
Steps to reproduce:
1) Create an invoice 1000 with a tax 21%
2) refund the invoice completely
3) Open the tax report: the base amount is -1000 instead of zero

Explanation:
The tax is using a different repartition line for invoice and for refund.
So, the tax report is getting 2 tax details: (-1000, -210) & (1000, 210) due to the groupby on tax_repartition_line.id.
The idea is to sum the base only once to avoid summing multiple times the same base amount when a tax has multiple repartition lines.
In our case, we need to consider the type of repartition line when grouping in order to count -1000+1000 = 0 as base amount.

closes odoo/enterprise#23599

X-original-commit: 3fc38dd96b11c61ddd21fdf1f869b2e17af82e3f
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Laurent Smet <las@odoo.com>