[IMP] project_sale_subscription: update dependencies
Before this commit, the dependencies of `project_sale_subscription`
module were `project` and `sale_subscription`. However,
`sale_subscription` need `sale_management`, it means `sale_project` will
also installed when we install this module.

For this reason, this commit updates the dependencies of this module by
replacing the `project` dependency by the `sale_project` to correctly
follow the order of the dependency modules in the project modules.

task-2710808