[IMP] project_sale_subscription: avoid counting subscription if no AA
Before this commit, when the `_compute_subscription_count` is trigger
for at least project, a `read_group` is called to count the number of
subscriptions linked to the project via the AA of this one. However, if
the project has no AA set, we could directly say no subscription is
linked to this project.

This commit updates the compute to directly set to 0
`subscription_count` field for all projects set in `self` if we don't have
at least an AA into a project contained into the `self`. By doing this,
we avoid an unecessary `read_group` onto `sale.subscription` model.

task-2710808