[FIX] web_map: make map view mobile friendly
Currently, when map view is default view for a action but it will
open the kanban as default view in mobile but map view is also
a mobile friendly so it should open the map view if default view
is map in mobile.

So in this commit, make map view mobile_fiendly so when map view
is default in action it should open map view in mobile.

task-2740041

Closes: #23656
X-original-commit: f31bf2b08d11d953c716bf5982194bfc97b64248
Part-of: odoo/enterprise#24808