[FIX] {test_}web_grid: fix web_grid_grouped when m2o is False
Before this commit, when the `sectionField` is a Many2one and in at least
one record we have False in this sectionField then we have a traceback
saying "Bool is not scriptable", the reason is when the m2o field used
to group the records is False on a record then
`section_field_value = group[section_field][0]` will crash because
`group[section_field] = False`.

This commit adds a check to be sure the `group[section_field]` is not
False before doing this `section_field_value = group[section_field][0]`.
Otherwise, we set the `section_field_value` to `group[section_field]`
like we do for the other type of fields used as `sectionField`.

Related #23096

closes odoo/enterprise#23627

X-original-commit: f03f88ccc0292f743a3adee1985a78dbd5f0f12c
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>