[FIX] account_reports: get_options_date_domain fix
Case:

In a financial report base on date range, with a line configured to display the
p&l carried forward as such.

"
<record id="report_line_ppl" model="account.financial.html.report.line">
    <field name="code">PPL</field>
    <field name="name">Profit and loss carried forward</field>
    <field name="sequence" eval="1"/>
    <field name="level" eval="3"/>
    <field name="domain" eval="[('account_id.user_type_id.internal_group', 'in', ['income', 'expense'])]"/>
    <field name="formulas">-sum</field>
    <field name="special_date_changer">to_beginning_of_period</field>
</record>
"

The line will be showing an amount at 0.
Displaying the P&L selecting a date in 2015 and one at the end of the last period (31/12 as of 24/01), we see
that we have a value that should be shown in the balance sheet amounting to 541.10.
Doing -sum of 541.10 on the above line is necessary in order to balance the balance sheet.

The code executed when calculating a line with a special_date_changer is setting it as such:
"
if self.special_date_changer == 'to_beginning_of_period' and date_from:
    date_tmp = fields.Date.from_string(self._context['date_from']) - relativedelta(days=1)
    date_to = date_tmp.strftime('%Y-%m-%d')
    date_from = False
"
Which seems correct.
But as it doesn't change the mode, we are in a range with only one value (date_to)

When calculating the dates domain in "get_options_date_domain", as no verification is done on the date_from
it will create a domain as such:
"
domain = [(date_field, '<=', options_date['date_to']), '|', (date_field, '>=', False),
          ('account_id.user_type_id.include_initial_balance', '=', True)]
"

The "(date_field, '>=', False)" would later be translated in SQL into "("account_move_line"."date" >= NULL)"
when calculating the sum for this line. This is not correct as it would always be False.

Thus, we aim to fix this by ensuring that, if we provide a range without any date_from, we do not append
the second part of the domain that is related to it.

closes odoo/enterprise#23723

X-original-commit: a15fe1dd802ee472eaaaa00645e067eb189a9fe9
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>