[FIX] l10n_mx_edi: fix compute of payment_method
before this commit, the default value was preventing any other
case in the computed method. So the default value was always the
only possible value for automatic invoicing

Now the compute method fallback on the previous default value when
no other value is found.

For automatic invoicing the payment method can come from the journal

closes odoo/enterprise#23680

X-original-commit: 5316b4d29b550f22ef5c2e1eba3b6c0b567795e7
Signed-off-by: Josse Colpaert <jco@odoo.com>