[FIX] account_reports: modifyin control_domain for LEX
How to reproduce the bug ?

- Install the accounting app
- Go to Profit and Loss report
- Add a new journal entry with these 2 items:
  - Cost of Goods Sold as account and a value in debit
  - Expenses as account and the same value in credit
- Save and post the entry
- Go back to the Profit and Loss report
- Enable the debug mode

What is the bug ?

In v15.0, there is a developer mode that allows the user to check if
there is an issue in the report. An error is detected if you add a
journal entry containing an item with the account Cost of Goods Sold.
Effectively, these items have their own section in the income section
while they are still taken into account in the expences computation.

opw-2691300

closes odoo/enterprise#25960

X-original-commit: fc00888a4798702cfa3e5215a8ea06bf20f638cc
Signed-off-by: William André (wan) <wan@odoo.com>