[FIX] project_forecast: fix the linked SOL should be visible in that template
When the user has the planning and Sales apps installed and created an open shift
(a shift without any resource set) and he published and sends the planning to his
employees, the employees can see the SOL linked to this shift. However, if the
user installs the Project app, the `project_forecast` module is installed to make
the bridge between the `Planning` and `Project` apps. This module overrides the template shown
to the employees to see the planning. This override erases the SOL field shown in this template.

Steps to reproduce:
Install Planning and Sales apps
Go to the planning app
Add a new shift and set an SOL
When we are in the Gantt view of the Planning app, clicks on the Publish button
and send the planning to an employee
Check the mail sent and go to the mail sent with the planning and click on the
link to redirect to the planning. Normally we see the SOL linked to the shift
Install the Project app
Go back to the planning URL sent to the employee

Actual Behaviour:
The SOL linked to the shift is no longer visible.

Expected Behaviour:
The SOL linked should be visible in that template.

closes odoo/enterprise#23705
task-2742510

closes odoo/enterprise#24334

X-original-commit: 087cb89749349eaf295f6df9384e056ce23296b2
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>