[FIX] l10n_co_reports: correct report missing rows
1. On a Colombian Company, create a Bill, add a CteFte tax to it, register a payment.
2. Go to Reports, Certificado de Retención en Fuente
3. Expan one Partner and print a preview

The original article or added information will not appear on the report.
This information is necessary for legal purposes in Colombia because this
is the legal basis on which the report is based.

This occur only when creating the pdf. HTML report is correct.
It seems wkhtmltopdf has issues with the row/div arrangment in this report
because the report will also look shrinked

opw-2684048

closes odoo/enterprise#23675

X-original-commit: f3187b5aa77df5395b17e16899b20f149778dd70
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>