[IMP] sign: remove get_signature route calls
When a signature or initial is required in sign, a RPC call is done to route
`get_signature`. This is bad for performance since we need to be doing http
calls everytime a wizard is opened. This commit removes the get_signature
route and now signatures are inserted in get_document_qweb_context as an
`auto_value` key in the sign_item_types dictionary. Also, `auto_field` keys
were replaced by `auto_value` to prevent confusion between the `auto_field`
field and the value that should be auto-filled in the document.

task-2735814

Part-of: odoo/enterprise#23653