[FIX] data_cleaning: commit after each batch for cron
data_cleaning_model._clean_records can create lots
of data_cleaning_records in one go. In case the cron
timeouts, the transaction is rolled back and no
record is written to the db.

As it is common for crons, it is better to split
the records to create in batches and commit after
each batch iteration. That way, if the transaction
gets rolled back, all the previously created records
are kept.

opw-2738529

closes odoo/enterprise#23739

X-original-commit: ca0a491823a19c7bdec13221ceee940073722982
Signed-off-by: Raphael Collet <rco@odoo.com>