[IMP] sign: prefetch fonts when signing document
In sign many NameAndSignature widgets are opened and closed since it is
expected that documents have multiple initial/signature fields. Everytime
the widget is opened, an rpc call to get_fonts is done, even if the request
has already been done before. This commit implements the prefetching of fonts
in PDFIframe and also the setting of fonts in the NameAndSignature override.

task-2735814

closes odoo/enterprise#23653

Related: odoo/odoo#83208
Signed-off-by: Arnaud Joset <arj@odoo.com>